+++ packages/core/src/translation/lv.ts
@@ -1,0 +1,20 @@
+import type { TranslationInterface } from '.';
+
+export const LatvianLanguage: TranslationInterface = {
+  write_a_message_placeholder: 'Rakstiet ziņojumu...',
+  your_issue_has_been_resolved: 'Jūsu problēma ir atrisināta!',
+  new_conversation: 'Jauna saruna',
+  welcome_screen_title: 'Laipni lūdzam mūsu atbalsta tērzēšanā',
+  welcome_screen_description:
+    'Mēs esam šeit, lai palīdzētu! Sāciet sarunu, un mēs jums atbildēsim pēc iespējas ātrāk.',
+  your_name_placeholder: 'Jūsu vārds',
+  your_email_placeholder: 'Jūsu e-pasta adrese',
+  start_chat_button: 'Sazināties ar atbalstu',
+  start_chat_button_loading: 'Savienošana...',
+  i_need_more_help: 'Man vajag vairāk palīdzības',
+  this_was_helpful: 'Tas bija noderīgi',
+  optional: 'Neobligāts',
+  no_conversations_yet: 'Pagaidām nav sarunu',
+  back_to_conversations: 'Atpakaļ pie sarunām',
+  closed_conversations: 'Slēgtās sarunas',
+};